Fixed transfer compilation fallback. It should fallback to the system temp directory if unable to write to the directory where the .bin file resides. It was doing that with the generated .java source file, but failing to properly fallback when renaming the compiled class file and then loading it. That should be resolved now. Was able to replicate issue and confirm that the class files show up with the proper file names in the system temp directory now, and are loading.
git-svn-id: file:///home/apertium/svn-mirror/apertium-sf-net/trunk/lttoolbox-java@27135 2c8c2297-680c-4500-a755-649f80e260dd

+++ src/org/apertium/transfer/TransferClassLoader.java
@@ -136,13 +136,22 @@ public class TransferClassLoader extends ClassLoader {
         fw.close();
 
         // Compile the source file
-        ApertiumTransferCompile.doMain(txFile, classFile, javaSource,
+        File actualClassFile = ApertiumTransferCompile.doMain(txFile, classFile, javaSource,
                 outputClass, false);
+        if(actualClassFile == null) {
+            /* We weren't able to get the compiled output file to rename
+             * to the desired classFile name, or to the same filename in
+             * the system temp directory. Throw an exception.
+             */
+            throw new ClassNotFoundException(
+                    "Unable to write compiled transfer class to specified " +
+                    "or system temp directories.");
+        }
 
         //Remove the java source file, as we don't need it anymore
         javaSource.delete();
         
         // Load and return the class file.
-        return tcl.loadClassFile(classFile);
+        return tcl.loadClassFile(actualClassFile);
     }
 }
